pyproject.toml
[tool.poetry]
name = "py-homelab"
version = "0.0.1"
description = ""
authors = ["Theo Paris <theoparisdesigns@gmail.com>"]

[tool.poetry.dependencies]
python = ">3.7"
json5 = "^0.9.5"
urwid = "^2.1.2"
Flask = "^1.1.2"
GitPython = "^3.1.9"
Flask-MongoAlchemy = "^0.7.2"
Flask-Bcrypt = "^0.7.1"
Flask-JWT-Extended = "^3.24.1"
throw-out-py = "^1.0.0"
python-dotenv = "^0.14.0"
flask-restx = "^0.2.0"

[tool.poetry.dev-dependencies]
pytest = "^4.6"
autopep8 = "^1.5.4"
throw-out-py = { path = "../throw-out-py", develop = true }

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
cli = "py_homelab.cli:main"
web = "py_homelab.web:main"
